POI 및 Swagger 추가

+++ src/main/java/com/deogicorgi/lotto/jpa/model/entity/Lotto.java
@@ -24,7 +24,7 @@ public class Lotto {
      * 회차
      */
     @Column(nullable = false)
-    private Long round;
+    private Integer round;
 
     /**
      * 추첨 일자
